option/Server/Kafka/requirements.txt
sinetstream
sinetstream-kafka
pypcapfile
